requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
PyMySQL==1.0.2
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
